1. Fixed how WHERE is generated 2. Fixed bug in find first bit set code 3. Make file changes of building kernel on SLES 11sp1

+++ tau.pj/include/bit.h
@@ -17,6 +17,6 @@
 #define lowBit(_x)	((_x) & ~((_x) - 1))
 #define highBit(_x)	(1 << ffsBit(_x))
 
-extern unsigned ffsBit(unsigned word);
+extern unsigned ffsBit (unsigned long word);
 
 #endif
